Removed Editor Code
Some other things have come up, and I'd like to reduce scope for the time being. Part of that scope reduction is to remove the editor. The game will now have a few pre-defined enemy classes that will be defined either in code, or via a simple text file the game will read.

The game is now going to be a hoard-mode survival game, rather than a mission based game. So an editor is un-necessary for the amount of data required for the game.