Physics Optimization
Moved collision check behaviour into it's own dedicated thread, which safely communicates a collision to WorldObjects. This has improved performance dramatically - when testing, we were running 256 collision checks per frame without a perf impact.

This can be optimized further by splitting the map up into grids, but given the game world is fairly small, we'll see how we get on with the current system. However, that is first on the list of optimizations if the framework was ever to be used elsewhere.